refactor: Merge `Testbed` and `RuntimeTestbed`
As the separation was not bringing much value and made existing code
more complex.

See #8201 for full motivation.
